streamlit/requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
langchain_openai
#faiss-cpu
azure-core
azure-search-documents
openai
azure-ai-vision-imageanalysis
pillow